[cleanup][client] Reduce repetition in admin client
### Motivation

In the admin client, each GET request handler implements its own
`InvocationCallback` which it implements anonymously at the callsites.
Jersey requires a concrete implementation to get around type erasure
since it uses reflection to identify the generic type. However, there
are few variants of the `InvocationCallback` implementations in the
admin client and most are identical - handling the response/exception
into a `CompletableFuture`.

### Modifications

Refactor the admin client GET handlers to be more DRY by providing a
`GetCallback` abstract implementation of `InvocationCallback` that
handles the response/exception into a `CompletableFuture` result.